Faster parquet DictEncoder
