+++ parquet/src/util/mod.rs
@@ -21,7 +21,6 @@ pub mod memory;
 pub mod bit_util;
 mod bit_packing;
 pub mod cursor;
-pub mod hash_util;
 #[cfg(any(test, feature = "test_common"))]
 pub(crate) mod test_common;
 pub(crate)mod page_util;
